% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dplexutilities.r
\name{CLR}
\alias{CLR}
\title{Perform center logratio transform on the input data}
\usage{
CLR(bar.table)

CLR(bar.table)
}
\arguments{
\item{bar.table}{formatted sample barcode table}
}
\value{
center logratio transformed bar.table as class Matrixarray

center logratio transformed bar.table as class Matrixarray
}
\description{
The CLR function provided is nearly identical to the one used in the Seurat package
except log1p() has been replaced with log(). The rational for this change is
that valid barcode counts are always positive and substantially greater than 0.

The CLR function provided is nearly identical to the one used in the Seurat package
except log1p() has been replaced with log(). The rational for this change is
that valid barcode counts are always positive and substantially greater than 0.
}
\examples{
n.bar.table <- CLR(bar.table)
n.bar.table <- CLR(bar.table)
}
