% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dplexdatatools.r
\name{cluster.Remove}
\alias{cluster.Remove}
\title{cluster.Remove
accepts the output of id_clusters() and returns a bar.UMAP removing all points 
in the specified cluster}
\usage{
cluster.Remove(cluster.res, bad.cluster)
}
\arguments{
\item{cluster.res}{output of id_clusters()}

\item{bad.cluster}{parameter defining the cluster to be removed}
}
\value{

}
\description{
cluster.Remove
accepts the output of id_clusters() and returns a bar.UMAP removing all points 
in the specified cluster
}
