% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dplexutilities.r
\name{data.format}
\alias{data.format}
\title{Format sample barcode count matrix}
\usage{
data.format(UMI_count_matrix)

data.format(UMI_count_matrix)
}
\arguments{
\item{UMI_count_matrix}{raw count output}
}
\value{
formatted count matrix where each column corresponds to a barcode and each row to a cell UMI

formatted count matrix where each column corresponds to a barcode and each row to a cell UMI
}
\description{
A convenience function equipped to handle the outputs of CITE-seq count or the MULTIseq Alignment Suite

A convenience function equipped to handle the outputs of CITE-seq count or the MULTIseq Alignment Suite
}
\details{
data.format() performs the following pipeline manipulations:
casts data input to a matrix using as.matrix()
transposes the input data if necessary
removes any uninformative columns attached during the alignment

data.format() performs the following pipeline manipulations:
casts data input to a matrix using as.matrix()
transposes the input data if necessary
removes any uninformative columns attached during the alignment
}
\examples{
f.bar.table <- data.format(bar.table)
f.bar.table <- data.format(bar.table)
}
