% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dplexdatatools.r
\name{estimateSNR}
\alias{estimateSNR}
\title{estimateSNR
TODO method needs to be improved! This is a prototype function}
\usage{
estimateSNR(bar.table)
}
\arguments{
\item{bar.table}{}
}
\value{

}
\description{
estimateSNR
TODO method needs to be improved! This is a prototype function
}
