% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REclassifycells.r
\name{findReclassCells}
\alias{findReclassCells}
\title{findReclassCells
Changes made to acquisition of extrema from original deMULTIplex}
\usage{
findReclassCells(barTable, neg.cells)
}
\arguments{
\item{barTable}{}

\item{neg.cells}{}
}
\value{

}
\description{
findReclassCells
Changes made to acquisition of extrema from original deMULTIplex
}
