% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dclassifycells.r
\name{findThresh}
\alias{findThresh}
\title{findThresh()
modified acquisition of extrema}
\usage{
findThresh(call.list)
}
\arguments{
\item{call.list}{}
}
\value{

}
\description{
findThresh()
modified acquisition of extrema
}
\examples{

}
