% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dclassifycells.r
\name{force_classifycells}
\alias{force_classifycells}
\title{force_classifycells()}
\usage{
force_classifycells(barTable, q, QA = 1)
}
\arguments{
\item{q}{}
}
\value{
bc_calls, a list of singlets, doublets, and negatives called for quantile q
}
\description{
Assumptions made by deMULTIplex break down when the number of barcodes is
small and if there is unequal representation of barcodes.
}
\details{
force_classifycells() provides an additional control branch to handle barcodes
which do not behave as assumed by deMULTIplex

forcing cell classification will increase the number of doublets falsely called as singlets
}
