% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dplexdatatools.r
\name{id_clusters}
\alias{id_clusters}
\title{id_clusters
provides several methods to identify UMAP clusters}
\usage{
id_clusters(bar.UMAP, method = "CLARA_medoids", nClusters = 10)
}
\arguments{
\item{method}{"gmm", "CLARA_medoids", or "kmeans"}

\item{nClusters}{}
}
\value{

}
\description{
id_clusters
provides several methods to identify UMAP clusters
}
