% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dplexutilities.r
\name{plotHashes}
\alias{plotHashes}
\title{Covenience function for visualizing output of barUMAP()}
\usage{
plotHashes(bar.UMAP, plt = TRUE)

plotHashes(bar.UMAP, plt = TRUE)
}
\arguments{
\item{bar.UMAP}{output matrix from barUMAP() function}

\item{plt}{logical, should plots be printed to std out. defaults to True.
Otherwise ggplot gobs are returned as a list.}
}
\value{
If plt is FALSE returns list of ggplot grobs otherwise no return value

If plt is FALSE returns list of ggplot grobs otherwise no return value
}
\description{
Covenience function for visualizing output of barUMAP()

Covenience function for visualizing output of barUMAP()
}
\examples{
plotHashes(bar.UMAP)
plot_list <- plotHashes(bar.UMAP, plt = FALSE)
plotHashes(bar.UMAP)
plot_list <- plotHashes(bar.UMAP, plt = FALSE)
}
