% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dclassifycells.r
\name{quant_first_thresh}
\alias{quant_first_thresh}
\title{quant_first_thresh
uses polynomial aproximation to get calculate derivative
TODO implement diff()}
\usage{
quant_first_thresh(model.x)
}
\arguments{
\item{model.x}{}
}
\value{
1st derivative zeroes
}
\description{
quant_first_thresh
uses polynomial aproximation to get calculate derivative
TODO implement diff()
}
