% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dclassifycells.r
\name{quant_second_thresh}
\alias{quant_second_thresh}
\title{quant_second_thresh
Uses polynomial approximation to calculate 2nd derivative
TODO diff(diff())}
\usage{
quant_second_thresh(model.x)
}
\arguments{
\item{model.x}{}
}
\value{
2nd derivative zeroes
}
\description{
quant_second_thresh
Uses polynomial approximation to calculate 2nd derivative
TODO diff(diff())
}
