% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dplexutilities.r
\name{signalNoise}
\alias{signalNoise}
\title{Calculate the signal to noise ratio for each cell in a formatted sample barcode table}
\usage{
signalNoise(bar.table)

signalNoise(bar.table)
}
\arguments{
\item{bar.table}{formatted sample barcode table}
}
\value{
a vector of signal to noise ratios

a vector of signal to noise ratios
}
\description{
Calculate the signal to noise ratio for each cell in a formatted sample barcode table

Calculate the signal to noise ratio for each cell in a formatted sample barcode table
}
\examples{
snr <- signalNoise(bar.table)
snr <- signalNoise(bar.table)
}
